[CWS] approve flag values above 31 in kernel space
flag_approver tests a value against a 64 bit mask by shifting a literal
`1`, which is an int, so every bit above 31 is lost. User space builds
the mask with `1 << (enum % 64)` on a u64 and does set those bits, so any
approved value of 32 or more produces a filter the kernel can never
match: the event is discarded before it reaches user space and the rule
that asked for it never fires.

For the BPF target clang performs the shift in 64 bits and then
truncates and sign extends it, so the mask is exactly zero for shifts of
32 and above, and 0xffffffff80000000 for a shift of 31, which
over-approves bits 31 to 63.

130 SECL constants across the six fields that use a flag approver land
in that window, among them:

    socket.domain == AF_VSOCK              go mask 0x0000010000000000
    bpf.cmd == BPF_LINK_DETACH             go mask 0x0000000400000000
    setsockopt.optname == SO_ATTACH_BPF    go mask 0x0004000000000000
    prctl.option == PR_SET_PTRACER
    prctl.option == PR_SET_NO_NEW_PRIVS
    connect.addr_family == AF_XDP

bpf and prctl have no other approver, so a discard there drops the event
outright. socket, setsockopt and connect fall back to their other
fields, but a rule that only constrains the one field installs no other
filter, so those fall through to a discard as well.

Shift a u64 instead, matching mask_has_event and the rest of the
codebase.

The accompanying unit test walks all 64 bits and checks that a mask with
a single bit set approves its own value and discards the neighbouring
one. It is split into a low and a high half so the failing section names
which part of the mask broke. Built against the int shift, the high half
is not merely wrong: the shift is undefined behaviour, and clang emits
the section with a size of zero.

+++ pkg/security/ebpf/c/include/helpers/approvers.h
@@ -274,7 +274,9 @@ static enum SYSCALL_STATE __attribute__((always_inline)) flag_approver (struct u
     if (filter == NULL || !filter->is_set) {
         return DISCARDED;
     }
-    if (((1 << (value % 64)) & filter->flags) > 0) {
+    // the shift has to be done on a u64: flags is a 64 bit mask and value % 64
+    // reaches 63, so an int shift would truncate every bit above 31.
+    if ((((u64)1 << (value % 64)) & filter->flags) > 0) {
         monitor_event_approved(type, FLAG_APPROVER_TYPE);
         return APPROVED;
     }
